compile netio on the fly for arm32/64 and x86_64

+++ src/libNetioHW/include/NetioHandler.h
@@ -44,7 +44,7 @@ public:
   std::vector<uint32_t> pushOut(uint64_t chn); // Push out every records from channel queue.
   SharedQueue& getQueue(uint64_t chn){ return m_pcqs[chn]; } // Access for elink's queue.
   size_t getNumOfChannels() { return m_activeChannels; } // Get the number of active channels.
-  bool isStable(uint64_t monitorID); // Returns the stability of elink's queue.
+  bool isStable(size_t monitorID); // Returns the stability of elink's queue.
   bool isAllStable(); // Returns the aggregated stability of the queues.
   void setFelixHost(std::string felixHost){m_felixHost=felixHost;}
   void setFelixRXPort(uint16_t felixRXPort){m_felixRXPort=felixRXPort;}
